Revert "Revert "Refactor mesh utilities into dedicated module""
