+++ analysis/__init__.py
@@ -9,6 +9,11 @@ from .flight_review import (
 from .summarise_runs import summarise_log
 from . import visualise_flight
 from .performance_plots import plot_performance
+from .mesh_utils import (
+    add_environment_mesh_to_plot,
+    extract_mesh_data,
+    apply_mesh_corrections,
+)
 
 __all__ = [
     "parse_log",
@@ -18,4 +23,7 @@ __all__ = [
     "summarise_log",
     "visualise_flight",
     "plot_performance",
+    "add_environment_mesh_to_plot",
+    "extract_mesh_data",
+    "apply_mesh_corrections",
 ]
